refactor(protocol): normalize test layout to one tests/ per capability
Stage 5 of the packages/protocol restructure. No public API change: all 443
exported symbols are byte-identical to dev, and no test changed meaning.

Six tests/ directories sat a level deeper than their capability's own --
intents/application/tests, opportunities/{discriminator,outcome,
negotiation-evidence}/tests, questions/{domain,ports}/tests -- so a capability
could have three. They merge into one per capability: nested 13 -> 7, total
27 -> 21.

The seven that remain are all under shared/, which has no capability root.
Merging those would produce a single 40-file directory mixing model config,
HyDE, schemas, and observability; one tests/ per shared module is the
consistent reading of the same rule there.

Also moves the two specs that sat outside any tests/ directory, and renames the
four .test.ts files to .spec.ts against 211 already using that suffix.

Removes all 12 test tsconfig.json files. They were not load-bearing: 9 of the 21
tests/ directories never had one, and the build, isolated runner, and tsc all
pass without them -- the build excludes tests by both **/tests/** and
**/*.spec.ts, and Bun does not read them. networks/tests/tsconfig.json had been
extending ../../tsconfig.json, which resolves to src/tsconfig.json and does not
exist; tsc reports TS5083 and silently falls back to defaults, dropping
esModuleInterop. It had been broken with nothing to notice.

Colocating every spec beside its subject was considered and rejected on
evidence: only 155 of 215 specs (72%) name a single source module. The other 60
are cross-module behaviour tests -- negotiation.continuation.spec.ts,
introducer-gating-lifecycle.spec.ts, negotiation.seat-rules.spec.ts -- and
filing those next to one arbitrary module they partly exercise would be worse
than leaving them grouped.

    directories  84 -> 78
    files       500 -> 488
    tests/ dirs  27 -> 21
    test tsconfigs 12 -> 0
    .test.ts       4 -> 0
    stray specs    2 -> 0

Verification:
- bun run build: clean
- bun run test (both halves): 2218 pass, 0 fail, 0 errors
- bun run eval:verify: all 14 suites
- bun run test:scripts: 193 pass
- public surface: 443 exported symbols byte-identical to dev
- runtime import cycles: 0
- bun run architecture:check, check:protocol-atlas: exit 0
- services/api tsc --noEmit: clean
- every entry in source-test-preload.ts and LIVE_MODEL_SPECS re-verified to
  resolve, and spec imports checked including mock.module() and dynamic
  import() paths, not just `from` clauses
