--- packages/protocol/src/networks/tests/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": { "noEmit": true },
-  "include": ["./**/*.spec.ts"]
-}
