+++ packages/protocol/src/opportunities/tests/discriminator.shadow.spec.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from "bun:test";
 
-import { runPoolDiscriminatorShadow } from "../discriminator.shadow.js";
-import type { EmbeddingGenerator } from "../../../shared/interfaces/embedder.interface.js";
-import type { MinedDiscriminator, PoolCandidate } from "../discriminator.types.js";
+import { runPoolDiscriminatorShadow } from "../discriminator/discriminator.shadow.js";
+import type { EmbeddingGenerator } from "../../shared/interfaces/embedder.interface.js";
+import type { MinedDiscriminator, PoolCandidate } from "../discriminator/discriminator.types.js";
 
 const candidates: PoolCandidate[] = Array.from({ length: 6 }, (_, i) => ({
   id: `c${i}`,
