--- packages/protocol/src/opportunities/tests/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "noEmit": true,
-    "composite": false,
-    "target": "ES2022",
-    "types": ["bun", "node"]
-  },
-  "include": ["**/*.spec.ts", "**/*.test.ts", "**/*.ts", "../**/*.ts"],
-  "exclude": ["node_modules"]
-}
