+++ packages/protocol/src/shared/agent/tests/tool.helpers.findPending.spec.ts
